% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleFree.R
\name{ScaleFree}
\alias{ScaleFree}
\title{Generate Scale free precision matrix}
\usage{
ScaleFree(p, powerparm, numadd, numT)
}
\arguments{
\item{p}{Dimension of network}

\item{powerparm}{Power parameter governing the preferential attachment model.  Small values produce random networks while large values lead to scale free networks.}

\item{numadd}{number of edges to add each iteration}

\item{numT}{number of timepoints}
}
\value{
A list with components
\item{Precision}{Estimated p by p inverse covariance matrix}
\item{Data}{Data matrix containing the simulated time series for each region (numT by p data matrix)}
}
\description{
Estimate a precision matrix with a scale free network structure and simulate multivariate normal data.
}
\examples{
fit<-ScaleFree(40,1.2,3,200)
fit$Precision
fit$Data
}
\references{
Csardi G, Nepusz T: The igraph software package for complex network research, InterJournal, Complex Systems 1695. 2006. \url{http://igraph.org}
}
